Move window.ts from settings/restore to settings/userSettings
